requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
asyncpg
openpyxl
opencv-python
pillow
numpy
python-multipart
